+++ lib/add-header-xua.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Add X-UA-Compatible header to all responses
+ *
+ * @package salcode/boostrap-genesis
+ */
+
+add_action( 'send_headers', 'bsg_add_header_xua' );
+
+/**
+ * Send X-UA-Compatible header
+ */
+function bsg_add_header_xua() {
+    header( 'X-UA-Compatible: IE=edge,chrome=1' );
+}