+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Bootstrap Genesis Theme Change Log
 disabled
 - Modify check for UberMenu plugin to additionally look for `UberMenu` class
 - Correct wrong Previous/Next Post Links Markup when using XHTML
+- Add header 'X-UA-Compatible: IE=edge,chrome=1' to all responses
 
 ## [0.8.2] - 2015-08-10
 - Update Bootstrap files to 3.3.5